Time: 0 ms (100.00%) | Memory: 19.5 MB (93.93%) - LeetSync

+++ 662-maximum-width-of-binary-tree/maximum-width-of-binary-tree.cpp
@@ -17,19 +17,19 @@ public:
             return 0;
         
         int ans = 0;
-        queue<pair<TreeNode*, long long>> q;
+        queue<pair<TreeNode*, int>> q;
 
         q.push({root, 0});
 
         while(!q.empty()){
 
-            long long mini = q.front().second;
+            int mini = q.front().second;
             int size = q.size();
-            long long low, high;
+            int low, high;
 
             for(int i = 0; i < size; i++){
 
-                long long nodeId = q.front().second - mini;
+                int nodeId = q.front().second - mini;
                 TreeNode* node = q.front().first;
                 q.pop();
 
@@ -39,12 +39,12 @@ public:
                     high = nodeId;
                 
                 if(node->left)
-                    q.push({node->left, 2*nodeId + 1});
+                    q.push({node->left, 2LL*nodeId + 1});
                 if(node->right)
-                    q.push({node->right, 2*nodeId + 2});
+                    q.push({node->right, 2LL*nodeId + 2});
             }
 
-            ans = max(ans, (int)(high - low  + 1));
+            ans = max(ans, (high - low  + 1));
         }
 
         return ans;
